Consolidate qoutationapis into UpbitQuotationApi
