--- client/retrofit/src/main/kotlin/com/jongmin/upbit/client/retrofit/quotation/api/UpbitMarketApi.kt
@@ -1,12 +1,0 @@
-package com.jongmin.upbit.client.retrofit.quotation.api
-
-import com.jongmin.upbit.client.retrofit.quotation.api.protocol.UpbitMarketResponse
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface UpbitMarketApi {
-
-    @GET("/v1/market/all")
-    fun getMarkets(@Query("isDetails") isDetails: Boolean? = false): Call<List<UpbitMarketResponse>>
-}
