+++ client/retrofit/src/main/kotlin/com/jongmin/upbit/client/retrofit/quotation/api/UpbitQuotationApi.kt
@@ -6,7 +6,7 @@ import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
 
-interface UpbitCandleApi {
+interface UpbitQuotationApi {
 
     @GET("/v1/candles/minutes/{unit}")
     fun getUpbitMinuteCandles(
@@ -37,4 +37,22 @@ interface UpbitCandleApi {
         @Query("to") to: String? = "yyyy-MM-dd HH:mm:ss",
         @Query("count") count: Int?
     ): Call<List<MonthCandleResponse>>
+
+    @GET("/v1/market/all")
+    fun getMarkets(@Query("isDetails") isDetails: Boolean? = false): Call<List<UpbitMarketResponse>>
+
+    @GET("/v1/orderbook")
+    fun getOrderbooks(@Query("markets") markets: String): Call<List<UpbitOrderbookResponse>>
+
+    @GET("/v1/ticker")
+    fun getCurrentTicker(@Query("markets") markets: String): Call<List<UpbitTickerResponse>>
+
+    @GET("v1/trades/ticks")
+    fun getTradeTicks(
+        @Query("market") market: String,
+        @Query("to") to: String? = "HHmmss",
+        @Query("count") count: Int? = null,
+        @Query("cursor") cursor: String? = null,
+        @Query("daysAgo") daysAgo: Int? = null
+    ): Call<List<UpbitTickResponse>>
 }
